Settings for a successful print.
This adds bridges to keep the MX stem in place on the bed during prints
and helps with adhesion. The bridges can easily be snipped off
afterwards.

The stem needs to extend all the way to the XY origin for this to work.

We also need less tolerance and the walls can be made a bit lighter.
The reduced tolerance and the increased diameter for the stem (probably
out of spec, but works just fine with my switches) allow the slicer to
produce a better result with the otherwise super thin wall. A 0.4 nozzle
can now produce a separate outside wall, adding integrity and increasing
print robustness.

Many failed build attempts went into this ...